Refactoring: move ranking fetch logic to server API side
